Improvements, removed workarounds
